feat(notifications): move to communication app
