+++ app/communication/admin.py
@@ -1,7 +1,6 @@
 from django.contrib import admin
 
-from app.communication.models.mail import Mail
-from app.communication.models.warning import Warning
+from app.communication.models import Mail, Notification, Warning
 
 
 class MailAdmin(admin.ModelAdmin):
@@ -13,4 +12,5 @@ class MailAdmin(admin.ModelAdmin):
 
 
 admin.site.register(Mail, MailAdmin)
+admin.site.register(Notification)
 admin.site.register(Warning)
