+++ app/communication/factories/notification_factory.py
@@ -2,7 +2,7 @@ import factory
 from factory.django import DjangoModelFactory
 
 from app.content.factories.user_factory import UserFactory
-from app.content.models.notification import Notification
+from app.communication.models.notification import Notification
 
 
 class NotificationFactory(DjangoModelFactory):
