+++ app/communication/models/__init__.py
@@ -1,2 +1,3 @@
 from app.communication.models.mail import Mail
+from app.communication.models.notification import Notification
 from app.communication.models.warning import Warning
