+++ app/communication/serializers/__init__.py
@@ -1,1 +1,5 @@
+from app.communication.serializers.notification import (
+    NotificationSerializer,
+    UpdateNotificationSerializer,
+)
 from app.communication.serializers.warning import WarningSerializer
