+++ app/communication/urls.py
@@ -1,10 +1,11 @@
 from django.conf.urls import include, url
 from rest_framework import routers
 
-from app.communication.views.warning import WarningViewSet
+from app.communication.views import NotificationViewSet, WarningViewSet
 
 router = routers.DefaultRouter()
 
+router.register("notifications", NotificationViewSet, basename="notification")
 router.register("warnings", WarningViewSet, basename="warning")
 
 urlpatterns = [url(r"", include(router.urls))]
