+++ app/communication/views/__init__.py
@@ -1,1 +1,2 @@
+from app.communication.views.notification import NotificationViewSet
 from app.communication.views.warning import WarningViewSet
