+++ app/content/admin/admin.py
@@ -10,7 +10,6 @@ admin.site.register(models.Event)
 admin.site.register(models.News)
 admin.site.register(models.Category)
 admin.site.register(models.Priority)
-admin.site.register(models.Notification)
 admin.site.register(models.Cheatsheet)
 admin.site.register(models.Badge)
 admin.site.register(models.UserBadge)
