+++ app/content/factories/__init__.py
@@ -4,7 +4,6 @@ from app.content.factories.user_factory import UserFactory
 from app.content.factories.registration_factory import RegistrationFactory
 from app.content.factories.cheatsheet_factory import CheatsheetFactory
 from app.content.factories.news_factory import NewsFactory
-from app.content.factories.notification_factory import NotificationFactory
 from app.content.factories.page_factory import PageFactory
 from app.content.factories.page_factory import ParentPageFactory
 from app.content.factories.short_link_factory import ShortLinkFactory
