+++ app/content/models/__init__.py
@@ -9,7 +9,6 @@ from app.content.models.registration import Registration
 from app.content.models.page import Page
 from app.content.models.prioritiy import Priority
 from app.content.models.short_link import ShortLink
-from app.content.models.notification import Notification
 from app.content.models.strike import (
     Strike,
     get_strike_description,
