+++ app/content/serializers/__init__.py
@@ -7,10 +7,6 @@ from app.content.serializers.event import (
     EventSerializer,
 )
 from app.content.serializers.news import NewsSerializer
-from app.content.serializers.notification import (
-    NotificationSerializer,
-    UpdateNotificationSerializer,
-)
 from app.content.serializers.page import (
     PageSerializer,
     PageTreeSerializer,
