+++ app/fixture.json
@@ -312,7 +312,7 @@
     }
 },
 {
-    "model": "content.notification",
+    "model": "communication.notification",
     "pk": 1,
     "fields": {
         "created_at": "2021-03-08T19:51:56.552Z",
@@ -324,7 +324,7 @@
     }
 },
 {
-    "model": "content.notification",
+    "model": "communication.notification",
     "pk": 2,
     "fields": {
         "created_at": "2021-03-20T12:23:12.092Z",
